requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asn1crypto==0.24.0
attrs==19.1.0
Babel==2.7.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
bs4==0.0.1
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
clarifai==2.0.21
conda==4.7.10
conda-package-handling==1.3.11
configparser==3.5.0
cryptography==2.7
cssselect==1.0.3
cssutils==1.0.2
defusedxml==0.6.0
Django==1.11.22
docutils==0.14
email-to==0.1.0
enum34==1.1.6
et-xmlfile==1.0.1
Faker==0.7.10
functools32==3.2.3.post2
future==0.15.2
futures==3.2.0
google-api-core==1.9.0
google-api-python-client==1.6.4
google-auth==1.6.3
google-auth-oauthlib==0.3.0
google-cloud-core==0.29.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleads==18.1.0
googleapis-common-protos==1.5.9
httplib2==0.13.0
idna==2.8
image-slicer==0.1.1
imagesize==1.1.0
ipaddress==1.0.22
isodate==0.6.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==2.5.1
libarchive-c==2.8
lxml==4.4.0
Markdown==3.1.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.16
numpy==1.16.4
oauth2client==4.1.3
oauthlib==3.0.1
openpyxl==2.6.2
pandas==0.24.2
Pillow==6.0.0
premailer==3.6.0
protobuf==3.8.0
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
PyMySQL==0.9.3
pyOpenSSL==19.0.0
PySocks==1.7.0
pytesseract==0.2.7
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rsa==4.0
ruamel-yaml==0.15.46
scipy==1.0.0
selenium==3.141.0
six==1.12.0
snowballstemmer==1.9.0
soupsieve==1.9.2
SpeechRecognition==3.6.3
Sphinx==1.5.5
suds-jurko==0.6
tqdm==4.32.1
uritemplate==3.0.0
urllib3==1.25.3
virtualenv==16.6.1
xmltodict==0.12.0
zeep==3.3.1
